Refactor: Split Neat.ts into focused modules (#1599)
Extract Neat.ts (1,312 lines) into focused modules:
- NeatEvolution.ts: Core evolution loop and result processing
- NeatScheduling.ts: Discovery and training task scheduling

Neat.ts is now a ~360-line facade that delegates to these modules.
