+++ src/NEAT/Neat.ts
@@ -1,50 +1,38 @@
-import { assert } from "@std/assert";
-import { blue } from "@std/fmt/colors";
-import { format } from "@std/fmt/duration";
+/**
+ * Neat.ts - Central class for the NEAT algorithm.
+ *
+ * Issue #1599: Refactored from 1,312 lines into a ~420-line facade.
+ * Responsibilities are delegated to focused modules in src/neat/:
+ * - NeatEvolution.ts   - Core evolution loop
+ * - NeatScheduling.ts  - Discovery and training task scheduling
+ */
+
 import { ensureDirSync } from "@std/fs";
-import { addTag, getTag, removeTag } from "@stsoftware/tags/mod";
 import { Creature } from "../Creature.ts";
 import { CreatureUtil } from "../architecture/CreatureUtils.ts";
 import { creatureValidate } from "../architecture/CreatureValidate.ts";
 import { DeDuplicator } from "../architecture/DeDuplicator.ts";
-import {
-  makeElitists,
-  sortCreaturesByScore,
-} from "../architecture/ElitismUtils.ts";
 import { Fitness } from "../architecture/Fitness.ts";
-import { calculate as calculateScore } from "../architecture/Score.ts";
 import { AdaptiveFineTuneTracker } from "../blackbox/AdaptiveFineTuneTracker.ts";
-import { fineTuneImprovement } from "../blackbox/FineTune.ts";
-import { FindTunePopulation } from "../blackbox/FineTunePopulation.ts";
 import { Breed } from "../breed/Breed.ts";
-import { ParallelBreeding } from "../breed/ParallelBreeding.ts";
 import { createNeatConfig, type NeatConfig } from "../config/NeatConfig.ts";
 import type { NeatOptions } from "../config/NeatOptions.ts";
-import type { TrainOptions } from "../config/TrainOptions.ts";
-
 import type {
   ResponseData,
   WorkerHandler,
 } from "../multithreading/workers/WorkerHandler.ts";
 import { WorkerPool } from "../multithreading/WorkerPool.ts";
-import { AddConnection } from "../mutate/AddConnection.ts";
+import type { CrisprInterface } from "../reconstruct/CRISPR.ts";
 import { Genus } from "./Genus.ts";
-import type { Approach } from "./LogApproach.ts";
 import { Mutator } from "./Mutator.ts";
-import { CRISPR, type CrisprInterface } from "../reconstruct/CRISPR.ts";
-import { simplify } from "../optimize/Simplify.ts";
-import { DiscoverStructure } from "../architecture/ErrorGuidedStructuralEvolution/DiscoverStructure.ts";
-import { isRustDiscoveryEnabled } from "../architecture/ErrorGuidedStructuralEvolution/RustDiscovery.ts";
-import { validateAfterDiscoveryOrThrow } from "../discovery/DiscoveryPostValidate.ts";
-import { calculateDiscoveryTimeout } from "../discovery/DiscoveryTimeout.ts";
 import { PlateauDetector } from "./PlateauDetector.ts";
-import {
-  type DiscoveryReplayDirResult,
-  DiscoveryReplayQueue,
-} from "./DiscoveryReplayQueue.ts";
+import { DiscoveryReplayQueue } from "./DiscoveryReplayQueue.ts";
 import { getLogger } from "../utils/Logger.ts";
 import { getRandomNumberGenerator } from "../utils/RandomNumberGenerator.ts";
-import { checkMemoryAndEvict, logMemoryUsage } from "./MemoryMonitor.ts";
+
+// Extracted modules
+import * as evolution from "./NeatEvolution.ts";
+import * as scheduling from "./NeatScheduling.ts";
 
 /**
  * NEAT (NeuroEvolution of Augmenting Topologies) implementation.
@@ -52,13 +40,6 @@ import { checkMemoryAndEvict, logMemoryUsage } from "./MemoryMonitor.ts";
  * This class implements the NEAT algorithm for evolving neural networks.
  * NEAT is a genetic algorithm that evolves both the topology and weights
  * of neural networks simultaneously.
- *
- * Key features:
- * - Population management and evolution
- * - Species-based selection and reproduction
- * - Mutation and crossover operations
- * - Fitness evaluation and scoring
- * - Multi-threaded training and discovery
  *
  * @example
  * ```ts
@@ -102,42 +83,43 @@ export class Neat {
   readonly workerPool: WorkerPool;
 
   /** Data directory for discovery replay (Issue #997) */
-  private dataDir?: string;
+  dataDir?: string;
 
   /** Count of critical-level evictions; used to throttle memory log noise (Issue #1565). */
-  private _memoryCriticalEvictionCount?: number;
+  memoryCriticalEvictionCount?: number;
 
-  /**
-   * Creates a new NEAT instance for evolving neural networks.
-   *
-   * @param input - Number of input neurons in the networks
-   * @param output - Number of output neurons in the networks
-   * @param options - Configuration options for the NEAT algorithm
-   * @param workers - Array of worker handlers for parallel processing
-   */
+  // Scheduling state - accessed by extracted scheduling/evolution modules
+  doNotStartMore = false;
+  private cleanUpDelayCount = 0;
+  additionalGenerationCount = 0;
+  private discoveryWaitGenerations = 0;
+  private maxDiscoveryWaitGenerations = 0;
+  trainingInProgress = new Map<string, Promise<void>>();
+  discoveryInProgress = new Map<string, Promise<void>>();
+  discoveryComplete: ResponseData[] = [];
+  trainingComplete: ResponseData[] = [];
+  alreadyScheduledMap = new Map<string, number>();
+  lastDiscoveryDurationMS = 0;
+  readonly MIN_DISCOVERY_TIME_MINUTES = 2;
+
   constructor(
     input: number,
     output: number,
     options: NeatOptions,
     workers: WorkerHandler[],
   ) {
-    this.input = input; // The input size of the networks
-    this.output = output; // The output size of the networks
+    this.input = input;
+    this.output = output;
     this.workers = workers;
-
-    // Custom cost functions are now handled by file path in WorkerHandler
-    // No registration needed in the main thread
 
     this.config = createNeatConfig(options);
 
-    // The fitness function to evaluate the networks
     this.fitness = new Fitness(
       this.workers,
       this.config.costOfGrowth,
       this.config.feedbackLoop,
     );
 
-    // Initialize the genomes
     this.population = [];
     this.config.creatures.forEach((c) => {
       const n = Creature.fromJSON(c);
@@ -149,45 +131,26 @@ export class Neat {
       : 0;
     this.CRISPRs = Neat.deepCloneAndShuffle(this.config.CRISPRs);
 
-    // Initialize plateau detector (Issue #1039)
     this.plateauDetector = new PlateauDetector(this.config.plateauDetection);
 
-    // Issue #1323: Initialize adaptive fine-tune population tracker
     this.fineTuneTracker = new AdaptiveFineTuneTracker(
       this.config.fineTunePopulation,
     );
 
-    // Initialize discovery replay queue (Issue #997)
     this.discoveryReplayQueue = new DiscoveryReplayQueue();
 
-    // Issue #1290: Initialise worker pool with work-stealing queues
     this.workerPool = new WorkerPool(this.workers);
   }
 
-  /**
-   * Set the data directory for discovery replay (Issue #997).
-   * This is needed because the Neat class doesn't have direct access to the
-   * data directory - it's typically passed to evolveDir in Creature.
-   *
-   * @param dataDir The data directory containing training data
-   */
   setDataDir(dataDir: string): void {
     this.dataDir = dataDir;
   }
 
-  /**
-   * Deep clones and shuffles an array using JSON serialization.
-   *
-   * @param arr - The array to clone and shuffle
-   * @returns A new shuffled array with deep-cloned elements
-   */
   static deepCloneAndShuffle<T>(arr: T[]): T[] {
     if (arr.length === 0) return [];
 
-    // Deep clone using JSON — can fail with non-serializable fields
     const cloned = JSON.parse(JSON.stringify(arr)) as T[];
 
-    // Shuffle (Fisher-Yates)
     const rng = getRandomNumberGenerator();
     for (let i = cloned.length - 1; i > 0; i--) {
       const j = Math.floor(rng.random() * (i + 1));
@@ -195,12 +158,6 @@ export class Neat {
     }
     return cloned;
   }
-
-  private doNotStartMore = false;
-  private cleanUpDelayCount = 0;
-  private additionalGenerationCount = 0;
-  private discoveryWaitGenerations = 0;
-  private maxDiscoveryWaitGenerations = 0;
 
   finishUp(
     iterations?: number,
@@ -222,29 +179,25 @@ export class Neat {
     }
 
     if (this.discoveryInProgress.size > 0) {
-      // Calculate max wait generations on first discovery wait
       if (this.maxDiscoveryWaitGenerations === 0) {
-        const DEFAULT_MAX_WAIT_MS = 5 * 60 * 1000; // 5 minutes
-        let maxWaitByIterations = 20; // Default fallback
-        let maxWaitByTime = 20; // Default fallback
+        const DEFAULT_MAX_WAIT_MS = 5 * 60 * 1000;
+        let maxWaitByIterations = 20;
+        let maxWaitByTime = 20;
 
         if (iterations) {
           maxWaitByIterations = Math.max(1, Math.floor(iterations * 0.5));
         }
 
-        // Calculate based on actual average time per generation if available
         if (startTimeMS && currentGeneration && currentGeneration > 0) {
           const elapsedMS = Date.now() - startTimeMS;
           const avgTimePerGeneration = elapsedMS / currentGeneration;
 
           if (avgTimePerGeneration > 0) {
-            // Calculate how many generations fit in 5 minutes
             const maxGenerationsIn5Minutes = Math.max(
               1,
               Math.floor(DEFAULT_MAX_WAIT_MS / avgTimePerGeneration),
             );
 
-            // If we have endTimeMS, also calculate based on remaining time
             if (endTimeMS) {
               const remainingTimeMS = endTimeMS - Date.now();
               if (remainingTimeMS > 0) {
@@ -255,7 +208,6 @@ export class Neat {
                 maxWaitByTime = Math.floor(estimatedGenerations * 0.5);
               }
             } else {
-              // Use the 5-minute limit if no endTimeMS
               maxWaitByTime = maxGenerationsIn5Minutes;
             }
 
@@ -266,10 +218,8 @@ export class Neat {
             );
           }
         } else if (endTimeMS) {
-          // Fallback to old estimation if we don't have timing data
           const remainingTimeMS = endTimeMS - Date.now();
           if (remainingTimeMS > 0) {
-            // Estimate generations based on remaining time (assume 1 generation per 30 seconds as rough estimate)
             const estimatedGenerations = Math.max(
               1,
               Math.floor(remainingTimeMS / 30000),
@@ -299,12 +249,11 @@ export class Neat {
           }`,
         );
 
-        // Clear stuck discoveries to allow evolution to complete
         this.discoveryInProgress.clear();
         this.discoveryWaitGenerations = 0;
         this.maxDiscoveryWaitGenerations = 0;
 
-        return false; // Continue to next iteration to check other conditions
+        return false;
       }
 
       const inProgressUUIDs = Array.from(this.discoveryInProgress.keys()).map(
@@ -342,408 +291,16 @@ export class Neat {
     return true;
   }
 
-  private trainingInProgress = new Map<string, Promise<void>>();
-  private discoveryInProgress = new Map<string, Promise<void>>();
-  discoveryComplete: ResponseData[] = [];
-  trainingComplete: ResponseData[] = [];
-
-  private alreadyScheduledMap = new Map<string, number>();
-  private lastDiscoveryDurationMS = 0;
-  private readonly MIN_DISCOVERY_TIME_MINUTES = 2;
+  // --- Delegated methods ---
 
   scheduleDiscovery(creature: Creature, timeOutMinutes: number) {
-    if (this.config.discoverySampleRate <= 0) {
-      return;
-    }
-
-    // Skip discovery if timeout is 0 or negative (discovery disabled)
-    if (
-      timeOutMinutes <= 0 || this.config.discoveryRecordTimeOutMinutes <= 0
-    ) {
-      return;
-    }
-
-    // Skip discovery if Rust library or GPU is not available
-    if (!isRustDiscoveryEnabled()) {
-      return;
-    }
-
-    if (this.doNotStartMore) return;
-
-    if (this.discoveryInProgress.size > 0) return;
-
-    const uuid = CreatureUtil.makeUUID(creature);
-
-    // Issue #1290: Use work-stealing pool for smarter worker selection
-    const w = this.workerPool.selectWorker();
-    if (!w) {
-      getLogger().warn("[Neat] No workers available for discovery");
-      return;
-    }
-
-    if (this.config.verbose) {
-      getLogger().info(
-        `Discovery ${
-          blue(uuid.substring(Math.max(0, uuid.length - 8)))
-        } scheduled`,
-      );
-    }
-
-    // Issue #1298: Adaptive discovery timeout based on creature complexity.
-    // Cap the recording-phase timeout so simple creatures don't wait unnecessarily
-    // long, while complex creatures are allowed more time.
-    const adaptiveTimeout = calculateDiscoveryTimeout({
-      neuronCount: creature.neurons.length,
-      synapseCount: creature.synapses.length,
-    });
-    const effectiveTimeout = Math.min(timeOutMinutes, adaptiveTimeout);
-
-    if (this.config.verbose) {
-      getLogger().info(
-        `[Neat] Adaptive discovery timeout: ${adaptiveTimeout.toFixed(2)}m ` +
-          `(neurons=${creature.neurons.length}, synapses=${creature.synapses.length}), ` +
-          `effective=${effectiveTimeout.toFixed(2)}m`,
-      );
-    }
-
-    // Issue #1538: Use lightweight frozen override instead of full createNeatConfig()
-    const discoveryConfig: NeatConfig = Object.freeze({
-      ...this.config,
-      discoveryRecordTimeOutMinutes: effectiveTimeout,
-    });
-
-    const taskStartTime = Date.now();
-    if (this.config.verbose) {
-      getLogger().info(
-        `[Neat] Discovery request sent to worker for ${
-          blue(uuid.substring(Math.max(0, uuid.length - 8)))
-        }`,
-      );
-    }
-
-    // Use internal timeout handling in DiscoverDirectory.ts instead of external wrapper
-    // This allows internal diagnostics to run and partial results to be returned
-    const discoveryPromise = w.discover(creature, discoveryConfig);
-
-    const p = discoveryPromise.then((r) => {
-      assert(r.discover, "No discovery found");
-
-      const discoveryDurationMS = Date.now() - taskStartTime;
-      this.lastDiscoveryDurationMS = discoveryDurationMS;
-
-      if (this.config.verbose) {
-        getLogger().info(
-          `[Neat] Discovery completed for ${
-            blue(uuid.substring(Math.max(0, uuid.length - 8)))
-          } after ${(discoveryDurationMS / 1000).toFixed(1)}s`,
-        );
-      }
-
-      // Issue #1020: Build the improved creature from the original creature.
-      // This ensures we apply discovery results to the creature that was analysed,
-      // not the current fittest (which may have evolved during discovery).
-      // We build ONE improved creature and add it directly to the population.
-      const improvedCreature = DiscoverStructure.addHelpfulSynapses(
-        uuid,
-        creature,
-        r.discover.addHelpfulSynapses,
-        this.config.discoveryFailureCacheDir,
-      );
-
-      if (improvedCreature) {
-        // Tag the improved creature to indicate it came from discovery
-        addTag(improvedCreature, "approach", "discovered");
-        addTag(improvedCreature, "discoveryID", uuid);
-
-        // Store the improved creature JSON in the response for direct population addition
-        r.discover.improvedCreature = improvedCreature.exportJSON();
-
-        if (this.config.verbose) {
-          getLogger().info(
-            `[Neat] Built improved creature from discovery ${
-              blue(uuid.substring(Math.max(0, uuid.length - 8)))
-            } with ${
-              r.discover.addHelpfulSynapses?.length ?? 0
-            } synapses added`,
-          );
-        }
-      }
-
-      this.discoveryComplete.push(r);
-
-      this.discoveryInProgress.delete(uuid);
-    }).catch((error) => {
-      getLogger().error(
-        `[Neat] Discovery failed for creature ${
-          uuid.substring(Math.max(0, uuid.length - 8))
-        } after ${((Date.now() - taskStartTime) / 1000).toFixed(1)}s:`,
-        error,
-      );
-
-      // Remove from discoveryInProgress to prevent infinite waiting
-      this.discoveryInProgress.delete(uuid);
-
-      // Add empty result to avoid breaking downstream logic
-      this.discoveryComplete.push({
-        taskID: 0,
-        duration: 0,
-        discover: {
-          ID: uuid,
-        },
-      });
-    });
-
-    this.discoveryInProgress.set(uuid, p);
+    scheduling.scheduleDiscovery(this, creature, timeOutMinutes);
   }
 
-  /**
-   * Issue #1150: Log a summary of the discovery replay result.
-   *
-   * This provides visibility into which candidates were evaluated during
-   * replay and which ones were successful. The summary is always logged
-   * when a replay completes, while detailed candidate information is
-   * only logged in verbose mode.
-   *
-   * @param result The discovery replay result to log
-   */
-  logReplaySummary(result: DiscoveryReplayDirResult): void {
-    const totalEvaluated = result.evaluatedSingles + result.evaluatedCombos;
-
-    // Build the summary parts
-    const parts: string[] = [];
-    parts.push(`evaluated: ${totalEvaluated}`);
-    if (result.evaluatedSingles > 0) {
-      parts.push(`singles: ${result.evaluatedSingles}`);
-    }
-    if (result.evaluatedCombos > 0) {
-      parts.push(`combos: ${result.evaluatedCombos}`);
-    }
-    if (result.pruned > 0) {
-      parts.push(`pruned: ${result.pruned}`);
-    }
-    if (result.skippedAlreadyApplied > 0) {
-      parts.push(`already-applied: ${result.skippedAlreadyApplied}`);
-    }
-    if (result.skippedNotApplicable > 0) {
-      parts.push(`not-applicable: ${result.skippedNotApplicable}`);
-    }
-    if (result.timedOut) {
-      parts.push("(timed out)");
-    }
-
-    // Log outcome
-    if (result.improvement) {
-      const changeType = result.improvement.changeType ?? "unknown";
-      const scoreDelta = result.improvement.scoreDelta?.toFixed(6) ?? "N/A";
-      getLogger().info(
-        `[Neat] Discovery replay: ${
-          blue(changeType)
-        } improved score by ${scoreDelta} (${parts.join(", ")})`,
-      );
-    } else {
-      getLogger().info(
-        `[Neat] Discovery replay: no improvement found (${parts.join(", ")})`,
-      );
-    }
-
-    // In verbose mode, log details of successful single candidates
-    if (this.config.verbose && result.evaluations) {
-      const successfulCandidates = result.evaluations.filter(
-        (e) => e.improved && e.kind !== "original",
-      );
-      if (successfulCandidates.length > 0) {
-        getLogger().info("[Neat] Successful replay candidates:");
-        for (const candidate of successfulCandidates) {
-          const kind = candidate.kind === "combo" ? "combo" : "single";
-          const changeType = candidate.changeType ?? "unknown";
-          const description = candidate.description ?? candidate.key ?? "";
-          const scoreDelta = candidate.scoreDelta?.toFixed(6) ?? "N/A";
-          getLogger().info(
-            `  - [${kind}] ${
-              blue(changeType)
-            }: ${description} (+${scoreDelta})`,
-          );
-        }
-      }
-    }
+  scheduleTraining(creature: Creature, trainingTimeOutMinutes: number) {
+    scheduling.scheduleTraining(this, creature, trainingTimeOutMinutes);
   }
 
-  /**
-   * Schedules training for a creature
-   * @param creature The creature to train
-   * @param trainingTimeOutMinutes The time out in minutes
-   */
-  scheduleTraining(
-    creature: Creature,
-    trainingTimeOutMinutes: number,
-  ) {
-    const uuid = CreatureUtil.makeUUID(creature);
-    if (this.trainingInProgress.has(uuid)) return;
-
-    if (this.alreadyScheduledMap.has(uuid)) {
-      if (!this.config.enableRepetitiveTraining) {
-        return;
-      }
-    }
-
-    this.alreadyScheduledMap.set(uuid, Date.now());
-    if (this.alreadyScheduledMap.size > 1000) {
-      // Clean up by removing old entries
-      const minuteAgo = Date.now() - 60_000;
-      for (const [key, value] of this.alreadyScheduledMap) {
-        if (value < minuteAgo) {
-          this.alreadyScheduledMap.delete(key);
-        }
-      }
-    }
-
-    // Issue #1290: Use work-stealing pool for smarter worker selection
-    const w = this.workerPool.selectWorker();
-    if (!w) {
-      getLogger().warn("[Neat] No workers available for training");
-      return;
-    }
-
-    if (this.config.verbose) {
-      getLogger().info(
-        `Training ${
-          blue(uuid.substring(Math.max(0, uuid.length - 8)))
-        } scheduled`,
-      );
-    }
-
-    const trainOptions: TrainOptions = {
-      log: this.config.log,
-      traceStore: this.config.traceStore,
-      iterations: 1,
-      targetError: this.config.targetError,
-      trainingSampleRate: this.config.trainingSampleRate,
-      disableRandomSamples: this.config.disableRandomSamples,
-      trainingTimeOutMinutes: trainingTimeOutMinutes,
-      batchSize: this.config.trainingBatchSize,
-      maximumBiasAdjustmentScale: this.config.maximumBiasAdjustmentScale,
-      maximumWeightAdjustmentScale: this.config.maximumWeightAdjustmentScale,
-      feedbackLoop: this.config.feedbackLoop,
-      sparseRatio: this.config.sparseRatio,
-      predictiveCoding: this.config.predictiveCoding,
-    };
-
-    const p = w.train(creature, trainOptions).then((r) => {
-      assert(r.train, "No train found");
-
-      const errorTx = getTag(creature, "error");
-      assert(errorTx, "No error tag found");
-
-      let trainingImprovement = true;
-      if (r.train.error > parseFloat(errorTx)) {
-        getLogger().warn(
-          `Training ${
-            blue(r.train.ID)
-          } caused a higher error of ${r.train.error} from ${errorTx}`,
-        );
-        trainingImprovement = false;
-      }
-      const trainedCreature = Creature.fromJSON(JSON.parse(r.train.creature));
-      trainedCreature.score = calculateScore(
-        trainedCreature,
-        r.train.error,
-        this.config.costOfGrowth,
-      );
-      const backtracked = fineTuneImprovement(
-        creature,
-        trainedCreature,
-        this.config.feedbackLoop,
-        1,
-        true,
-        this.config.quantumStep,
-      );
-      const forward = fineTuneImprovement(
-        creature,
-        trainedCreature,
-        this.config.feedbackLoop,
-        1,
-        false,
-        this.config.quantumStep,
-      );
-
-      if (backtracked.length > 0 || forward.length > 0) {
-        const untrainedError = getTag(trainedCreature, "untrained-error");
-        if (backtracked.length > 0) {
-          const backtrackedCreature = backtracked[0].exportJSON();
-          addTag(backtrackedCreature, "trainID", r.train.ID);
-          addTag(backtrackedCreature, "trainVariant", "overshot");
-          if (untrainedError) {
-            addTag(backtrackedCreature, "untrained-error", untrainedError);
-          }
-          r.train.backtracked = JSON.stringify(backtrackedCreature);
-        }
-        if (forward.length > 0) {
-          const forwardCreature = forward[0].exportJSON();
-          addTag(forwardCreature, "trainID", r.train.ID);
-          addTag(forwardCreature, "trainVariant", "undershot");
-          if (untrainedError) {
-            addTag(forwardCreature, "untrained-error", untrainedError);
-          }
-          r.train.forward = JSON.stringify(forwardCreature);
-        }
-      } else if (trainingImprovement === false) {
-        getLogger().warn(
-          `Training ${
-            blue(r.train.ID)
-          } caused a higher error of ${r.train.error} from ${errorTx} and no tuning`,
-        );
-      }
-
-      this.trainingComplete.push(r);
-
-      this.trainingInProgress.delete(uuid);
-
-      if (this.config.traceStore) {
-        if (r.train.trace) {
-          const traceNetwork = Creature.fromJSON(
-            JSON.parse(r.train.trace),
-          );
-          CreatureUtil.makeUUID(traceNetwork);
-          ensureDirSync(this.config.traceStore);
-          Deno.writeTextFileSync(
-            `${this.config.traceStore}/${traceNetwork.uuid}.json`,
-            JSON.stringify(traceNetwork.traceJSON(), null, 1),
-          );
-        }
-      }
-    }).catch((error) => {
-      getLogger().error(
-        `Training failed for creature ${
-          uuid.substring(Math.max(0, uuid.length - 8))
-        }:`,
-        error,
-      );
-
-      // Remove from trainingInProgress to prevent infinite waiting
-      this.trainingInProgress.delete(uuid);
-
-      // Add empty result to avoid breaking downstream logic
-      this.trainingComplete.push({
-        taskID: 0,
-        duration: 0,
-        train: {
-          ID: uuid,
-          creature: JSON.stringify(creature.exportJSON()),
-          error: Number.POSITIVE_INFINITY,
-          trace: "",
-        },
-      });
-    });
-
-    this.trainingInProgress.set(uuid, p);
-  }
-
-  /**
-   * Evaluates, selects, breeds and mutates population.
-   *
-   * @param previousFittest - The fittest creature from the previous generation
-   * @returns Evolution results including fittest creature, average score, and plateau status
-   */
   async evolve(
     previousFittest?: Creature,
   ): Promise<{
@@ -756,509 +313,7 @@ export class Neat {
       mutationMultiplier: number;
     };
   }> {
-    this.additionalGenerationCount--;
-
-    await this.fitness.calculate(this.population);
-    sortCreaturesByScore(this.population);
-
-    const genus = new Genus();
-
-    this.writeScores(
-      this.population,
-    );
-
-    // The population is already sorted in the desired order
-    for (const creature of this.population) {
-      assert(creature.uuid, "UUID missing");
-      assert(creature.score, "Score missing");
-      assert(
-        Number.isFinite(creature.score),
-        `Score: ${creature.score} is not finite`,
-      );
-      genus.addCreature(creature);
-    }
-    if (this.population.length === 0) {
-      getLogger().warn("All creatures died, using zombies");
-    }
-
-    const numberOfElitists = this.config.elitism > 1
-      ? this.config.elitism
-      : previousFittest
-      ? this.config.elitism
-      : 2;
-
-    /* Elitism: we need at least 2 on the first run */
-    const results = makeElitists(
-      this.population,
-      numberOfElitists,
-      this.config.verbose,
-    );
-    const elitists = results.elitists;
-
-    let tmpFittest = elitists[0];
-
-    assert(tmpFittest.uuid, "Fittest creature has no UUID");
-    assert(tmpFittest.score, "No fittest creature score found");
-    let previousFittestUUID = "NONE";
-    if (previousFittest) {
-      previousFittestUUID = previousFittest.uuid ?? "NONE";
-      assert(previousFittest.score, "No previous fittest creature score found");
-      assert(previousFittest.uuid, "Previous fittest creature has no UUID");
-      assert(
-        previousFittest.score <= tmpFittest.score,
-        "Previous fittest has a higher score than fittest",
-      );
-      if (previousFittest.score === tmpFittest.score) {
-        if (previousFittest.uuid !== tmpFittest.uuid) {
-          getLogger().info(
-            `Fittest creature ${
-              tmpFittest.uuid.substring(0, 8)
-            } has the same score as previous fittest ${
-              previousFittest.uuid.substring(0, 8)
-            } reuse previous fittest.`,
-          );
-        }
-        tmpFittest = previousFittest;
-      }
-    }
-
-    // Issue #1025: Use shallowClone() instead of fromJSON(exportJSON())
-    // for significant performance improvement (3-4x faster for large creatures).
-    const fittest = tmpFittest.shallowClone(); // Make a copy so it's not mutated.
-
-    fittest.score = tmpFittest.score;
-    assert(fittest.score, "No fittest score found");
-
-    // Issue #1039: Record fitness for plateau detection
-    this.plateauDetector.recordFitness(fittest.score);
-
-    // Issue #1323: Record whether fine-tuning produced the new fittest
-    if (previousFittest) {
-      const approach = getTag(tmpFittest, "approach");
-      const fineTuneImproved = approach === "fine" &&
-        tmpFittest.uuid !== previousFittest.uuid;
-      this.fineTuneTracker.recordOutcome(fineTuneImproved);
-    }
-
-    addTag(fittest, "score", fittest.score.toString());
-
-    const error = getTag(fittest, "error");
-    assert(error, "No error tag found");
-    let trainingTimeOutMinutes = 0;
-    if (this.endTimeTS) {
-      const diff = this.endTimeTS - Date.now();
-      trainingTimeOutMinutes = Math.round(diff / 60_000);
-
-      if (trainingTimeOutMinutes < 1) {
-        trainingTimeOutMinutes = -1;
-      }
-    }
-
-    if (previousFittestUUID !== fittest.uuid) {
-      const simplified = simplify(fittest);
-
-      if (simplified) {
-        elitists.push(simplified);
-      }
-
-      if (trainingTimeOutMinutes !== -1) {
-        // Estimate if we have enough time for discovery
-        const estimatedDiscoveryMinutes = this.lastDiscoveryDurationMS > 0
-          ? Math.ceil(this.lastDiscoveryDurationMS / 60000) + 1 // Add 1 min buffer
-          : this.MIN_DISCOVERY_TIME_MINUTES;
-
-        if (trainingTimeOutMinutes >= estimatedDiscoveryMinutes) {
-          this.scheduleDiscovery(fittest, trainingTimeOutMinutes);
-        } else {
-          if (this.config.verbose) {
-            getLogger().info(
-              `Skipping discovery: insufficient time (${trainingTimeOutMinutes}m remaining, ${estimatedDiscoveryMinutes}m estimated)`,
-            );
-          }
-        }
-      }
-
-      // Issue #997: Schedule background replay of cached discoveries against the new fittest.
-      // This allows learnings from previous discovery runs to be applied when evolution
-      // progresses. The replay runs in a non-blocking manner, one at a time.
-      // Issue #1113: Pass remaining time to replay so it can timeout gracefully.
-      if (this.dataDir && this.config.discoveryCacheDir) {
-        this.discoveryReplayQueue.scheduleReplay(
-          fittest,
-          this.dataDir,
-          this.config,
-          trainingTimeOutMinutes > 0 ? trainingTimeOutMinutes : undefined,
-        );
-      }
-    }
-
-    if (trainingTimeOutMinutes !== -1) { // If not timed out already
-      for (
-        let i = 0;
-        i < results.elitists.length;
-        i++
-      ) {
-        const n = results.elitists[i];
-
-        if (
-          this.doNotStartMore === false &&
-          this.trainingInProgress.size < this.config.trainPerGen &&
-          Number.isFinite(n.score)
-        ) {
-          this.scheduleTraining(
-            n,
-            trainingTimeOutMinutes,
-          );
-        }
-      }
-    }
-
-    const dnaPopulation = [];
-
-    if (this.CRISPRs.length) {
-      const crispr = new CRISPR(fittest);
-      while (this.CRISPRs.length > 0) {
-        const dna = this.CRISPRs.pop()!;
-
-        const enhanced = crispr.cleaveDNA(dna);
-        if (enhanced.uuid !== fittest.uuid) {
-          dnaPopulation.push(enhanced);
-          break;
-        }
-      }
-    }
-    const newPopulation = [];
-    if (
-      elitists.length > 0
-    ) {
-      const n = elitists[0];
-      // Issue #1040: Use shallowClone() instead of fromJSON(exportJSON())
-      // for better performance - avoids expensive JSON serialisation/deserialisation
-      const creativeThinking = n.shallowClone();
-      delete creativeThinking.memetic;
-      // Clear score to match behaviour of previous JSON clone approach
-      // (exportJSON doesn't export score, so fromJSON wouldn't have it)
-      delete creativeThinking.score;
-      const weightScale = 1 / Math.max(creativeThinking.synapses.length, 1);
-      const addConnection = new AddConnection(creativeThinking);
-      for (let i = 0; i < this.config.creativeThinkingConnectionCount; i++) {
-        addConnection.mutate(
-          getRandomNumberGenerator().random() < this.config.focusRate
-            ? this.config.focusList
-            : undefined,
-          {
-            weightScale: weightScale,
-          },
-        );
-      }
-
-      assert(!creativeThinking.memetic);
-      newPopulation.push(creativeThinking);
-    }
-
-    const ftp = new FindTunePopulation(this, this.fineTuneTracker);
-    const fineTunedPopulation = ftp.make(
-      fittest,
-      previousFittest,
-      genus,
-    );
-
-    const newPopSize = this.config.populationSize -
-      elitists.length -
-      dnaPopulation.length -
-      this.trainingComplete.length -
-      this.discoveryComplete.length -
-      fineTunedPopulation.length - 1 -
-      newPopulation.length;
-
-    // Issue #1026: Use parallel breeding for improved performance
-    // Pass workers for true parallelism across multiple CPU cores
-    const parallelBreeding = new ParallelBreeding(
-      genus,
-      this.config,
-      this.workers,
-    );
-    const offspringBatch = await parallelBreeding.breedBatch(newPopSize);
-    newPopulation.push(...offspringBatch);
-
-    // Keep breed instance for DeDuplicator usage later
-    const breed = new Breed(genus, this.config);
-
-    // Issue #1039: Apply plateau stagnation response - increase mutation rate
-    // Issue #1538: Use lightweight frozen override instead of full createNeatConfig()
-    const mutationMultiplier = this.plateauDetector.getMutationMultiplier();
-    let mutatorConfig: NeatConfig = this.config;
-    if (mutationMultiplier > 1.0) {
-      // Create a modified config with increased mutation rate
-      const adjustedMutationRate = Math.min(
-        this.config.mutationRate * mutationMultiplier,
-        1.0, // Cap at 100%
-      );
-      mutatorConfig = Object.freeze({
-        ...this.config,
-        mutationRate: adjustedMutationRate,
-      });
-      if (this.config.verbose) {
-        getLogger().info(
-          `[Plateau] Stagnation detected - mutation rate increased from ` +
-            `${(this.config.mutationRate * 100).toFixed(1)}% to ` +
-            `${(adjustedMutationRate * 100).toFixed(1)}% ` +
-            `(${this.plateauDetector.getGenerationsOnPlateau()} generations on plateau)`,
-        );
-      }
-    }
-
-    const mutator = new Mutator(mutatorConfig);
-    // Replace the old population with the new population
-    mutator.mutate(newPopulation);
-
-    // Issue #1099: Single-pass de-duplication - create DeDuplicator but defer perform()
-    // until after all population sources are combined. This reduces overhead from
-    // running de-duplication twice (after breeding and again after combining sources).
-    // The previous two-pass approach (Issue #1014) caught duplicates early but
-    // redundantly checked newPopulation creatures again in the final pass.
-    const deDuplicator = new DeDuplicator(breed, mutator);
-
-    const trainedPopulation: Creature[] = [];
-
-    for (let i = this.trainingComplete.length; i--;) {
-      const r = this.trainingComplete[i];
-      assert(r.train, "No train found");
-      // Skip failed training results (those with non-finite errors like Infinity)
-      if (!Number.isFinite(r.train.error)) {
-        continue;
-      }
-
-      const json = JSON.parse(r.train.creature);
-      if (this.config.verbose) {
-        getLogger().info(
-          `Training ${blue(r.train.ID)} completed ${
-            r.duration
-              ? "after " + format(r.duration, { ignoreZero: true })
-              : ""
-          }`,
-        );
-      }
-
-      addTag(json, "approach", "trained" as Approach);
-      delete json.memetic;
-      removeTag(json, "approach-logged");
-      addTag(json, "trainID", r.train.ID);
-      addTag(json, "trained", "YES");
-
-      trainedPopulation.push(Creature.fromJSON(json, this.config.debug));
-      if (r.train.backtracked) {
-        fineTunedPopulation.push(
-          Creature.fromJSON(JSON.parse(r.train.backtracked), this.config.debug),
-        );
-      }
-      if (r.train.forward) {
-        fineTunedPopulation.push(
-          Creature.fromJSON(JSON.parse(r.train.forward), this.config.debug),
-        );
-      }
-      const compactJSON = r.train.compact
-        ? JSON.parse(r.train.compact)
-        : undefined;
-
-      if (compactJSON) {
-        if (this.config.verbose) {
-          getLogger().info(
-            `Training ${blue(r.train.ID)} compacted`,
-          );
-        }
-
-        addTag(compactJSON, "approach", "compact" as Approach);
-        delete compactJSON.memetic;
-        removeTag(compactJSON, "approach-logged");
-        addTag(compactJSON, "trainID", r.train.ID);
-        addTag(compactJSON, "trained", "YES");
-
-        trainedPopulation.push(
-          Creature.fromJSON(compactJSON, this.config.debug),
-        );
-      }
-
-      // Immediately clear large objects to help GC
-      // @ts-ignore - clearing to help GC
-      r.train.creature = null;
-      // @ts-ignore - clearing to help GC
-      r.train.trace = null;
-      // @ts-ignore - clearing to help GC
-      r.train.compact = null;
-      // @ts-ignore - clearing to help GC
-      r.train.backtracked = null;
-      // @ts-ignore - clearing to help GC
-      r.train.forward = null;
-    }
-    this.trainingComplete.length = 0;
-
-    // Issue #1020: Process discovery results by adding discovered creatures directly.
-    // The improved creature was already built from the original creature in scheduleDiscovery(),
-    // so we just add it to the population without applying changes to the current fittest.
-    // This ensures evolution continues without being blocked by discovery results.
-    for (let i = this.discoveryComplete.length; i--;) {
-      const r = this.discoveryComplete[i];
-      assert(r.discover, "No discovery found");
-
-      // Issue #1020: Add the pre-built improved creature directly to the population.
-      // The creature was built in scheduleDiscovery() from the original creature,
-      // not the current fittest. This allows discoveries to be applied without
-      // blocking evolution or racing with population changes.
-      if (r.discover.improvedCreature) {
-        const discoveredCreature = Creature.fromJSON(
-          r.discover.improvedCreature,
-        );
-        CreatureUtil.makeUUID(discoveredCreature);
-
-        validateAfterDiscoveryOrThrow({
-          baseCreature: fittest,
-          discoveredCreature: discoveredCreature,
-          discoveryID: r.discover.ID,
-          operation: "discovered-creature-addition",
-          feedbackLoop: this.config.feedbackLoop,
-        });
-
-        trainedPopulation.push(discoveredCreature);
-
-        if (this.config.verbose) {
-          getLogger().info(
-            `[Neat] Added discovered creature ${
-              blue(discoveredCreature.uuid?.substring(0, 8) ?? "unknown")
-            } to population (from discovery ${
-              blue(
-                r.discover.ID.substring(Math.max(0, r.discover.ID.length - 8)),
-              )
-            })`,
-          );
-        }
-      }
-
-      // Immediately clear large objects to help GC
-      // @ts-ignore - clearing to help GC
-      r.discover.addHelpfulSynapses = null;
-      // @ts-ignore - clearing to help GC
-      r.discover.removeHarmfulSynapse = null;
-      // @ts-ignore - clearing to help GC
-      r.discover.candidateSquashes = null;
-      // @ts-ignore - clearing to help GC
-      r.discover.improvedCreature = null;
-    }
-    this.discoveryComplete.length = 0;
-
-    // Issue #997: Process completed background replay results and add improved
-    // creatures to the population. These are creatures that have had cached
-    // discovery improvements re-applied from previous evolution runs.
-    // Issue #1150: Log detailed information about successful candidates
-    const replayResults = this.discoveryReplayQueue.getCompletedResults();
-    for (const result of replayResults) {
-      // Issue #1150: Log replay summary showing which candidates were evaluated
-      // and their outcomes. This provides visibility into the replay process
-      // even when no improvement is found.
-      this.logReplaySummary(result);
-
-      if (result.improvement?.creature) {
-        // The creature field is typed as `unknown` in DiscoveryReplayDirResult
-        // but is actually a CreatureExport from exportJSON()
-        const replayedCreature = Creature.fromJSON(
-          result.improvement.creature as Parameters<
-            typeof Creature.fromJSON
-          >[0],
-        );
-        CreatureUtil.makeUUID(replayedCreature);
-
-        // Tag the creature to indicate it came from discovery replay
-        addTag(replayedCreature, "approach", "discovery-replay");
-
-        validateAfterDiscoveryOrThrow({
-          baseCreature: fittest,
-          discoveredCreature: replayedCreature,
-          discoveryID: result.improvement.key ?? "replay",
-          operation: "discovery-replay-addition",
-          feedbackLoop: this.config.feedbackLoop,
-        });
-
-        trainedPopulation.push(replayedCreature);
-
-        if (this.config.verbose) {
-          getLogger().info(
-            `[Neat] Added replayed creature ${
-              blue(replayedCreature.uuid?.substring(0, 8) ?? "unknown")
-            } to population (score improvement: ${
-              result.improvement.scoreDelta?.toFixed(4) ?? "N/A"
-            })`,
-          );
-        }
-      }
-    }
-    this.discoveryReplayQueue.clearCompletedResults();
-
-    /** make sure we do at least one more loop */
-    if (trainedPopulation.length > 0 && this.additionalGenerationCount <= 0) {
-      this.additionalGenerationCount = 1;
-    }
-
-    // Issue #1568: Save reference to old population before replacement
-    // so we can dispose creatures that are not carried forward.
-    const oldPopulation = this.population;
-
-    this.population = [
-      ...elitists,
-      ...trainedPopulation,
-      ...fineTunedPopulation,
-      ...newPopulation,
-      ...dnaPopulation,
-    ]; // Keep pseudo sorted.
-
-    // Issue #1099: Single-pass de-duplication on the combined population
-    // This catches duplicates both within each source and between different
-    // population sources (elitists, trainedPopulation, fineTunedPopulation,
-    // newPopulation, dnaPopulation) in a single operation, reducing overhead
-    // compared to the previous two-pass approach.
-    deDuplicator.perform(this.population);
-
-    // Issue #1568: Explicitly dispose old population creatures that were not
-    // carried forward. This frees WASM heap allocations deterministically
-    // rather than waiting for the FinalizationRegistry / GC to clean up.
-    // Elitists are carried forward into the new population, so we skip those.
-    const carriedForward = new Set(this.population);
-    for (const creature of oldPopulation) {
-      if (!carriedForward.has(creature)) {
-        creature.dispose();
-      }
-    }
-
-    // Issue #1565: Proactive heap memory monitoring and graduated cache eviction.
-    // Checked once per generation after old population disposal.
-    const memoryResult = checkMemoryAndEvict(
-      this.config.memory,
-      getLogger(),
-    );
-    // Only log when we took action (evicted), to avoid flooding when heap is normal.
-    // When stuck at critical, throttle to every 10th eviction to reduce log noise.
-    if (memoryResult.evicted) {
-      if (memoryResult.pressureLevel !== "critical") {
-        logMemoryUsage(memoryResult, getLogger());
-      } else {
-        this._memoryCriticalEvictionCount =
-          (this._memoryCriticalEvictionCount ?? 0) + 1;
-        if (this._memoryCriticalEvictionCount % 10 === 1) {
-          logMemoryUsage(memoryResult, getLogger());
-        }
-      }
-    }
-
-    return {
-      fittest: fittest,
-      averageScore: results.averageScore,
-      // Issue #1039: Include plateau detection status in evolution results
-      plateau: {
-        onPlateau: this.plateauDetector.isOnPlateau(),
-        generationsOnPlateau: this.plateauDetector.getGenerationsOnPlateau(),
-        improvementRate: this.plateauDetector.getImprovementRate(),
-        mutationMultiplier: this.plateauDetector.getMutationMultiplier(),
-      },
-    };
+    return evolution.evolve(this, previousFittest);
   }
 
   writeScores(creatures: Creature[]) {
@@ -1280,9 +335,6 @@ export class Neat {
     }
   }
 
-  /**
-   * Create the initial pool of genomes
-   */
   populatePopulation(creature: Creature) {
     if (this.config.debug) {
       creatureValidate(creature);
@@ -1299,7 +351,6 @@ export class Neat {
 
     const genus = new Genus();
 
-    // The population is already sorted in the desired order
     for (const creature of this.population) {
       CreatureUtil.makeUUID(creature);
       genus.addCreature(creature);
